requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pandas==1.5.3
pillow==10.3.0
PyQt5==5.15.10
spyder-kernels==2.5.13
xlrd==2.0.1
xlwt==1.3.0